requirements.txt
pydantic[email]
pinecone-client
pypdf
tiktoken
fpdf
unstructured
textract==1.6.3
gunicorn==20.1.0
google-cloud-storage==2.18.2
fastapi==0.116.2
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-cloud-secret-manager==2.20.2
langchain-google-genai==2.0.1
faiss-cpu==1.9.0
docx2txt==0.8
python-multipart==0.0.12
alembic==1.13.3
vertexai
google-cloud-secret-manager 
google-cloud-aiplatform
docx
asyncpg==0.30.0
sqlmodel==0.0.27
pydantic-settings==2.10.1
bcrypt==4.0.1
passlib==1.7.4
python-dotenv==1.1.1
pinecone==7.3.0
sentence-transformers==5.1.0
PyPDF2==3.0.1
python-docx==1.2.0
pandas==2.3.2
openpyxl==3.1.5
google-generativeai==0.8.5
aiofiles==25.1.0
aiosmtplib
google-genai
firebase-admin
google-cloud-kms==3.7.0
